backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "altair>=6.0.0",
    "asyncpg>=0.31.0",
    "bcrypt<4.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.124.4",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "langchain>=1.2.0",
    "langchain-openai>=1.1.4",
    "litellm>=1.80.10",
    "minio>=7.2.20",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "plotly>=6.5.0",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.21",
    "redis>=7.1.0",
    "structlog>=24.4.0",
    "tenacity>=9.1.2",
    "tiktoken>=0.12.0",
    "uvicorn>=0.38.0",
    "daytona-sdk>=0.127.0",
    "openpyxl>=3.1.0",
    "nest-asyncio>=1.6.0",
    "Pillow>=11.0.0",
    "rich>=13.0.0",
    "matplotlib>=3.10.0",
    "seaborn>=0.13.0",
    "scipy>=1.15.0",
    "scikit-learn>=1.6.0",
    "kaleido>=1.2.0",
]
